--- .changeset/evil-hands-fry.md
@@ -1,5 +1,0 @@
----
-"@comet/cms-admin": patch
----
-
-Mark `FinalFormToggleButtonGroup` as deprecated. Use `ToggleButtonGroupField` instead.
